api4: break out license and config from system
